+++ crates/ruff/src/rules/flake8_pyi/rules/string_or_bytes_too_long.rs
